Divide la aplicación en modulos
